HUDManager class added
Text in the top center of the screen will now display the name of the last valid target hit!
It will only stay readable for a few seconds (set by a variable and executed through co-routines) before it's removed from the HUD. Keeping it present all the time would not look good.